Session manager implemented

+++ src/main/java/com/finalproject/finalproject/controller/CategoryController.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -14,14 +15,18 @@ public class CategoryController extends CustomExceptionHandler {
 
     @Autowired
     CategoryService categoryService;
+    @Autowired
+    SessionManager sessionManager;
 
-    @PostMapping("/category/{categoryName}")
-    public ResponseEntity<CategoryDTO> addCategoryToDB(@PathVariable("categoryName") CategoryNameDTO categoryName){
+    @PostMapping("/category")
+    public ResponseEntity<CategoryDTO> addCategoryToDB(@RequestParam("categoryName") CategoryNameDTO categoryName, HttpServletRequest request){
+        sessionManager.verifyAdmin(request);
         return ResponseEntity.ok(categoryService.addToDB(categoryName));
     }
-    @DeleteMapping("/category/{categoryName}")
-    public ResponseEntity<CategoryDTO> deleteCategoryFromDB(@PathVariable("categoryName") CategoryNameDTO categoryName) {
-          return ResponseEntity.ok(categoryService.deleteFromDB(categoryName));
+    @DeleteMapping("/category")
+    public ResponseEntity<CategoryDTO> deleteCategoryFromDB(@RequestParam("categoryName") CategoryNameDTO categoryName, HttpServletRequest request) {
+        sessionManager.verifyAdmin(request);
+        return ResponseEntity.ok(categoryService.deleteFromDB(categoryName));
     }
     @GetMapping("/category/all")
     public ResponseEntity<List<CategoryDTO>> getAll(){
